express-remix-frbarbre/django/requirements.txt
django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
pymongo==4.6.1
python-dotenv==1.0.0